instant-property-feature-detector-updated/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1
python-multipart==0.0.9
Pillow==10.3.0
opencv-python==4.10.0.84
ultralytics==8.3.6
